refactor: isolate Telegram response transport
